+++ module/main.c
@@ -1023,6 +1023,7 @@ int main(void) {
     flash_get_cal(&scene_state.cal);
     ss_update_param_scale(&scene_state);
     ss_update_in_scale(&scene_state);
+    ss_update_fader_scale_all(&scene_state);
 
     // load preset from flash
     preset_select = flash_last_saved_scene();