+++ src/ops/fader.h
@@ -4,6 +4,14 @@
 #include "ops/op.h"
 
 extern const tele_op_t op_FADER;
+extern const tele_op_t op_FADER_SCALE;
+extern const tele_op_t op_FADER_CAL_MIN;
+extern const tele_op_t op_FADER_CAL_MAX;
+extern const tele_op_t op_FADER_CAL_RESET;
 extern const tele_op_t op_FB;
+extern const tele_op_t op_FB_S;
+extern const tele_op_t op_FB_C_MIN;
+extern const tele_op_t op_FB_C_MAX;
+extern const tele_op_t op_FB_C_R;
 
 #endif